Reduce number of partial files to two
