+++ src/System.Collections.Immutable/src/System.Collections.Immutable.csproj
@@ -40,14 +40,7 @@
     <Compile Include="System\Collections\Immutable\ImmutableArray_1.Builder.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableArray_1.Enumerator.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableArray_1.cs" />
-    <Compile Include="System\Collections\Immutable\ImmutableArray_1.CollectionInterfaces.cs" />
-    <Compile Include="System\Collections\Immutable\ImmutableArray_1.IImmutableList.cs" />
-    <Compile Include="System\Collections\Immutable\ImmutableArray_1.IndexOf.cs" />
-    <Compile Include="System\Collections\Immutable\ImmutableArray_1.Insert.cs" />
-    <Compile Include="System\Collections\Immutable\ImmutableArray_1.OfType.cs" />
-    <Compile Include="System\Collections\Immutable\ImmutableArray_1.Remove.cs" />
-    <Compile Include="System\Collections\Immutable\ImmutableArray_1.Replace.cs" />
-    <Compile Include="System\Collections\Immutable\ImmutableArray_1.Sort.cs" />
+    <Compile Include="System\Collections\Immutable\ImmutableArray_1.Minimal.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableDictionary.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableDictionary_2.Builder.cs" />
     <Compile Include="System\Collections\Immutable\ImmutableDictionary_2.Comparers.cs" />
